Add type information to ctypes functions.
`argtypes` and `restype` are set for all functions.

Also `UtpContext` and `UtpSocket` incomplete types were removed. These
were causing trouble for some reason. The `POINTER(UtpSocket)` that we
received in the `on_accept` segfaulted when passed to `utp_write`
because it wasn't the pointer to the real socket. Passing the poiner
around as an integer, or as was decided in this case as a `c_void_p`
fixes the problem.